refactor(AFS-3): Rename PropertyType enum to plural and Properties.Property namespace to Properties.Properties to avoid class/namespace conflicts.
